# Backend/src/main/resources/openapi/device-manager-openapi.yaml
openapi: 3.0.3
info:
  title: Device Manager API
  version: 1.0.0
  description: >
    REST API for managing groups, users, menus, applications,
    wallpapers and themes in the Device Manager system.
  contact:
    name: Milán Buha
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/groups:
    get:
      tags:
        - Groups
      summary: List all groups
      operationId: getAllGroups
      responses:
        '200':
          description: Groups returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupSummaryResponse'

    post:
      tags:
        - Groups
      summary: Create a new group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupSummaryResponse'
  /api/groups/{groupId}/users:
    get:
      tags:
        - Users
      summary: List users in a group
      operationId: getUsersByGroup
      parameters:
        - name: groupId
          in: path
          required: true
          description: Unique identifier of the group
          schema:
            type: string
      responses:
        '200':
          description: Users returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummaryResponse'

    post:
      tags:
        - Users
      summary: Create a new user in a group
      operationId: createUser
      parameters:
        - name: groupId
          in: path
          required: true
          description: Unique identifier of the group
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
  /api/groups/{groupId}/users/{userId}:
    delete:
      tags:
        - Users
      summary: Delete a user from a group
      operationId: deleteUser
      parameters:
        - name: groupId
          in: path
          required: true
          description: Unique identifier of the group
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
  /api/groups/{groupId}/users/{userId}/login:
    post:
      tags:
        - Users
      summary: Authenticate a user
      operationId: loginUser
      parameters:
        - name: groupId
          in: path
          required: true
          description: Unique identifier of the group
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
  /api/groups/{groupId}/access-validation:
    post:
      tags:
        - Groups
      summary: Validate group access code
      operationId: validateAccessCode
      parameters:
        - name: groupId
          in: path
          required: true
          description: Unique identifier of the group
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateGroupAccessRequest'
      responses:
        '200':
          description: Access code validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupAccessValidationResponse'
...
  /api/users/{userId}/wallpaper:
    put:
      tags:
        - Customization
      summary: Select wallpaper for user
      operationId: selectWallpaper
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectWallpaperRequest'
      responses:
        '204':
          description: Wallpaper selected successfully
  /api/users/{userId}/theme:
    put:
      tags:
        - Customization
      summary: Select theme for user
      operationId: selectTheme
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectThemeRequest'
      responses:
        '204':
          description: Theme selected successfully
components:
  schemas:
    CreateGroupRequest:
      type: object
      description: Request body for creating a new group with its initial admin user
      required:
        - groupName
        - accessCode
        - adminUserName
        - adminPassword
      properties:
        groupName:
          type: string
          description: Name of the group
          example: Buha Family
        accessCode:
          type: string
          description: Access code used to enter the group
          example: asd
        adminUserName:
          type: string
          description: Name of the initial admin user
          example: Buha
        adminPassword:
          type: string
          description: Password of the initial admin user
          example: asd
    CreateUserRequest:
      type: object
      description: Request body for creating a new user inside a group
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: Name of the new user
          example: Niki
        password:
          type: string
          description: Password of the new user
          example: asd
...
    UserLoginRequest:
      type: object
      description: Request body for authenticating a user
      required:
        - password
      properties:
        password:
          type: string
          description: User password
          example: asd
    ValidateGroupAccessRequest:
      type: object
      description: Request body for validating a group's access code
      required:
        - accessCode
      properties:
        accessCode:
          type: string
          description: Access code of the group
          example: asd
    SelectWallpaperRequest:
      type: object
      description: Request body for selecting a wallpaper for a user
      required:
        - wallpaperId
      properties:
        wallpaperId:
          type: string
          description: Identifier of the selected wallpaper
          example: wallpaper-mountain
    SelectThemeRequest:
      type: object
      description: Request body for selecting a theme for a user
      required:
        - themeId
      properties:
        themeId:
          type: string
          description: Identifier of the selected theme
          example: theme-dark
    GroupSummaryResponse:
      type: object
      description: Short summary of a group
      properties:
        id:
          type: string
          description: Unique identifier of the group
          example: group-buha-family
        name:
          type: string
          description: Name of the group
          example: Buha Family
    UserSummaryResponse:
      type: object
      description: Short summary of a user
      properties:
        id:
          type: string
          description: Unique identifier of the user
          example: user-buha-admin
        name:
          type: string
          description: Name of the user
          example: Buha
        role:
          type: string
          description: Role of the user
          example: ADMIN
    GroupAccessValidationResponse:
      type: object
      description: Result of validating a group's access code
      properties:
        valid:
          type: boolean
          description: Whether the provided access code is valid
          example: true
        message:
          type: string
          description: Validation result message
          example: Az access code helyes.
...